+++ CHANGELOG.md
@@ -3,6 +3,13 @@ tf_hachef CHANGELOG
 
 This file is used to list changes made in each version of the tf_hachef Terraform plan.
 
+v0.2.0 (2016-08-08)
+-------------------
+- Overhaul on code (nearly complete re-write)
+- Updated syntax for (most) Terraform 0.7.0 constructs
+- NOTE: Leaving `template` in place of `data` source due to `count` absence on `data` source
+- Removed a number of files
+
 v0.1.1 (2016-07-15)
 -------------------
 - Documentation work